+++ resources/js/app.js
@@ -5,13 +5,8 @@ require('selectize')
 const jQueryBridget = require('jquery-bridget')
 const Isotope = require('isotope-layout')
 const InfiniteScroll = require('infinite-scroll')
-const Flickity = require('flickity');
-
-Flickity.setJQuery( $ );
-jQueryBridget('flickity', Flickity, $)
 
 jQueryBridget('isotope', Isotope, $)
-
 jQueryBridget('infiniteScroll', InfiniteScroll, $)
 InfiniteScroll.imagesLoaded = require('imagesloaded')
 
@@ -36,6 +31,10 @@ Vue.use(Lang, { fallback: 'sk' })
 Vue.component('filter-sort-by', require('./components/filter/SortBy.vue').default);
 Vue.component('filter-checkbox', require('./components/filter/Checkbox.vue').default);
 Vue.component('filter-custom-select', require('./components/filter/CustomSelect.vue').default);
+Vue.component('flickity', require('./components/Flickity.vue').default);
+Vue.component('tabs-controller', require('./components/TabsController.vue').default);
+Vue.component('tab', require('./components/Tab.vue').default);
+Vue.component('tab-panel', require('./components/TabPanel.vue').default);
 Vue.component('user-collections-link', require('./components/user-collections/Link.vue').default);
 Vue.component('user-collections-nav-link', require('./components/user-collections/NavLink.vue').default);
 Vue.component('user-collections-favourite-button', require('./components/user-collections/FavouriteButton.vue').default);