+++ resources/js/components/Flickity.vue
@@ -1,0 +1,18 @@
+<template>
+    <div>
+        <slot />
+    </div>
+</template>
+
+<script>
+import Flickity from "flickity"
+
+export default {
+    props: {
+        options: Object
+    },
+    mounted() {
+        new Flickity(this.$el, this.options)
+    },
+}
+</script>