+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - filter catalog by curator and fix harvesting of curator data
 - attributes "format", "deliveryPoint" and "note" to JIRA request in order process
+- basic Tailwind set-up
+- Tabs and Flickity Vue components
 
 ### Fixed
 - calculation of isFree for artworks with authorities without death_year